Conflicts:
	.gitignore
	admin/environment.xml
	admin/index.php
	admin/renderer.php
	admin/settings/appearance.php
	admin/settings/courses.php
	admin/settings/plugins.php
	admin/settings/security.php
	admin/settings/server.php
	admin/settings/subsystems.php
	admin/settings/top.php
	admin/settings/users.php
	admin/tool/phpunit/cli/init.php
	admin/tool/phpunit/cli/util.php
	admin/tool/unittest/lang/en/tool_unittest.php
	blocks/completionstatus/details.php
	calendar/renderer.php
	cohort/assign.php
	cohort/edit.php
	cohort/edit_form.php
	cohort/index.php
	cohort/lib.php
	course/category.php
	course/completion.php
	course/completion_form.php
	course/edit.php
	course/edit_form.php
	course/format/topics/format.php
	course/format/weeks/format.php
	course/format/weeks/lib.php
	course/index.php
	course/lib.php
	course/search.php
	enrol/category/db/events.php
	enrol/category/locallib.php
	enrol/cohort/lang/en/enrol_cohort.php
	enrol/cohort/locallib.php
	enrol/manual/lib.php
	grade/grading/simpletest/testlib.php
	group/index.php
	group/members.php
	lang/en/admin.php
	lang/en/completion.php
	lang/en/enrol.php
	lang/en/repository.php
	lib/clilib.php
	lib/completion/completion_completion.php
	lib/completion/completion_criteria.php
	lib/completion/completion_criteria_completion.php
	lib/completion/completion_criteria_course.php
	lib/completion/completion_criteria_date.php
	lib/completion/completion_criteria_grade.php
	lib/completion/cron.php
	lib/completionlib.php
	lib/db/install.xml
	lib/db/upgrade.php
	lib/dml/moodle_database.php
	lib/dml/simpletest/testdml.php
	lib/editor/tinymce/tiny_mce/3.4.9/plugins/moodlemedia/editor_plugin.js
	lib/editor/tinymce/tiny_mce/3.4.9/plugins/moodlemedia/editor_plugin_src.js
	lib/editor/tinymce/tiny_mce/3.4.9/themes/advanced/skins/o2k7/dialog.css
	lib/formslib.php
	lib/installlib.php
	lib/messagelib.php
	lib/moodlelib.php
	lib/navigationlib.php
	lib/outputrenderers.php
	lib/pagelib.php
	lib/phpunit/bootstraplib.php
	lib/phpunit/classes/advanced_testcase.php
	lib/phpunit/classes/data_generator.php
	lib/phpunit/classes/database_driver_testcase.php
	lib/phpunit/classes/hint_resultprinter.php
	lib/phpunit/classes/module_generator.php
	lib/phpunit/classes/util.php
	lib/phpunit/lib.php
	lib/pluginlib.php
	lib/sessionlib.php
	lib/setup.php
	lib/simpletest/testoutputlib.php
	lib/simpletest/testrss.php
	lib/tablelib.php
	lib/upgradelib.php
	lib/weblib.php
	lib/yui/3.4.1/build/cssbase-context/base-context-min.css
	lib/yui/3.4.1/build/cssbase-context/base-context.css
	lib/yui/3.4.1/build/cssbase-context/cssbase-context-min.css
	lib/yui/3.4.1/build/cssbase-context/cssbase-context.css
	lib/yui/3.4.1/build/cssbase/base-min.css
	lib/yui/3.4.1/build/cssbase/base.css
	lib/yui/3.4.1/build/cssbase/cssbase-min.css
	lib/yui/3.4.1/build/cssbase/cssbase.css
	lib/yui/3.4.1/build/cssreset/cssreset-min.css
	lib/yui/3.4.1/build/cssreset/cssreset.css
	lib/yui/3.4.1/build/cssreset/reset-min.css
	lib/yui/3.4.1/build/cssreset/reset.css
	login/confirm.php
	login/signup.php
	message/defaultoutputs.php
	mod/book/README.md
	mod/book/backup/moodle1/lib.php
	mod/book/backup/moodle2/backup_book_activity_task.class.php
	mod/book/backup/moodle2/backup_book_settingslib.php
	mod/book/backup/moodle2/backup_book_stepslib.php
	mod/book/backup/moodle2/restore_book_activity_task.class.php
	mod/book/backup/moodle2/restore_book_stepslib.php
	mod/book/db/access.php
	mod/book/db/install.xml
	mod/book/db/log.php
	mod/book/db/subplugins.php
	mod/book/db/upgrade.php
	mod/book/db/upgradelib.php
	mod/book/delete.php
	mod/book/edit.php
	mod/book/edit_form.php
	mod/book/index.php
	mod/book/lang/en/book.php
	mod/book/lib.php
	mod/book/locallib.php
	mod/book/mod_form.php
	mod/book/move.php
	mod/book/pix/icon.png
	mod/book/settings.php
	mod/book/show.php
	mod/book/styles.css
	mod/book/tool/exportimscp/db/access.php
	mod/book/tool/exportimscp/index.php
	mod/book/tool/exportimscp/lang/en/booktool_exportimscp.php
	mod/book/tool/exportimscp/lib.php
	mod/book/tool/exportimscp/locallib.php
	mod/book/tool/exportimscp/version.php
	mod/book/tool/importhtml/db/access.php
	mod/book/tool/importhtml/import_form.php
	mod/book/tool/importhtml/index.php
	mod/book/tool/importhtml/lang/en/booktool_importhtml.php
	mod/book/tool/importhtml/lib.php
	mod/book/tool/importhtml/locallib.php
	mod/book/tool/importhtml/version.php
	mod/book/tool/print/db/access.php
	mod/book/tool/print/index.php
	mod/book/tool/print/lang/en/booktool_print.php
	mod/book/tool/print/lib.php
	mod/book/tool/print/locallib.php
	mod/book/tool/print/version.php
	mod/book/version.php
	mod/book/view.php
	mod/feedback/db/install.xml
	mod/feedback/db/upgrade.php
	mod/scorm/db/install.xml
	mod/scorm/db/upgrade.php
	mod/scorm/lang/en/scorm.php
	mod/scorm/lib.php
	mod/scorm/locallib.php
	mod/scorm/mod_form.php
	mod/scorm/player.php
	mod/scorm/report/graphs/graph.php
	mod/scorm/report/graphs/version.php
	mod/scorm/styles.css
	mod/scorm/version.php
	mod/scorm/view.js
	mod/scorm/view.php
	phpunit.xml.dist
	report/completion/index.php
	report/completion/lang/en/report_completion.php
	report/completion/user.php
	theme/afterburner/config.php
	theme/afterburner/layout/default.php
	theme/afterburner/pix/menu/ab-arrowover.png
	theme/afterburner/pix/menu/nav-arrowover-right.png
	theme/afterburner/style/afterburner_calendar.css
	theme/afterburner/style/afterburner_dock.css
	theme/afterburner/style/afterburner_menu.css
	theme/afterburner/style/afterburner_styles.css
	theme/afterburner/version.php
	theme/anomaly/config.php
	theme/anomaly/renderers.php
	theme/anomaly/style/base.css
	theme/anomaly/style/dock.css
	theme/anomaly/style/general.css
	theme/anomaly/version.php
	theme/arialist/config.php
	theme/arialist/style/core.css
	theme/arialist/style/pagelayout.css
	theme/arialist/version.php
	theme/base/config.php
	theme/base/style/core.css
	theme/base/style/course.css
	theme/base/style/dock.css
	theme/base/style/filemanager.css
	theme/binarius/version.php
	theme/boxxie/config.php
	theme/boxxie/version.php
	theme/brick/version.php
	theme/canvas/config.php
	theme/canvas/layout/frontpage.php
	theme/canvas/layout/general.php
	theme/canvas/layout/report.php
	theme/canvas/style/blocks.css
	theme/canvas/style/core.css
	theme/canvas/style/mods.css
	theme/canvas/version.php
	theme/formal_white/config.php
	theme/formal_white/db/install.php
	theme/formal_white/db/upgrade.php
	theme/formal_white/lang/en/theme_formal_white.php
	theme/formal_white/layout/embedded.php
	theme/formal_white/layout/frontpage.php
	theme/formal_white/layout/general.php
	theme/formal_white/lib.php
	theme/formal_white/settings.php
	theme/formal_white/style/core.css
	theme/formal_white/style/course.css
	theme/formal_white/style/formal_white.css
	theme/formal_white/style/frame.css
	theme/formal_white/style/menu.css
	theme/formal_white/style/pagelayout.css
	theme/formal_white/style/quiz.css
	theme/formal_white/version.php
	theme/formfactor/config.php
	theme/formfactor/version.php
	theme/fusion/style/core.css
	theme/fusion/version.php
	theme/leatherbound/config.php
	theme/leatherbound/version.php
	theme/magazine/config.php
	theme/magazine/style/core.css
	theme/magazine/version.php
	theme/mymobile/renderers.php
	theme/mymobile/style/jmobile11_rtl.css
	theme/nimble/version.php
	theme/nonzero/version.php
	theme/overlay/layout/general.php
	theme/overlay/version.php
	theme/serenity/version.php
	theme/sky_high/style/core.css
	theme/sky_high/style/report.css
	theme/sky_high/version.php
	theme/splash/config.php
	theme/splash/lang/en/theme_splash.php
	theme/splash/layout/general.php
	theme/splash/layout/report.php
	theme/splash/lib.php
	theme/splash/settings.php
	theme/splash/style/blue.css
	theme/splash/style/editor.css
	theme/splash/style/green.css
	theme/splash/style/orange.css
	theme/splash/version.php
	theme/standard/pix/menu/ab-arrowover.png
	user/filesedit.php
	user/filesedit_form.php
	user/profile/field/menu/field.class.php
	user/profile/field/text/field.class.php
	version.php

+++ group/groupings.php
@@ -20,7 +20,7 @@
  *
  * @copyright 1999 Martin Dougiamas  http://dougiamas.com
  * @license http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
- * @package groups
+ * @package core_group
  */
 
 require_once '../config.php';
@@ -35,7 +35,7 @@ if (!$course = $DB->get_record('course', array('id'=>$courseid))) {
 }
 
 require_login($course);
-$context = get_context_instance(CONTEXT_COURSE, $course->id);
+$context = context_course::instance($course->id);
 require_capability('moodle/course:managegroups', $context);
 
 $strgrouping     = get_string('grouping', 'group');
@@ -67,6 +67,7 @@ echo $OUTPUT->heading($strgroupings);
 
 $data = array();
 if ($groupings = $DB->get_records('groupings', array('courseid'=>$course->id), 'name')) {
+    $canchangeidnumber = has_capability('moodle/course:changeidnumber', $context);
     foreach($groupings as $grouping) {
         $line = array();
         $line[0] = format_string($grouping->name);
@@ -84,8 +85,13 @@ if ($groupings = $DB->get_records('groupings', array('courseid'=>$course->id), '
 
         $buttons  = "<a title=\"$stredit\" href=\"grouping.php?id=$grouping->id\"><img".
                     " src=\"" . $OUTPUT->pix_url('t/edit') . "\" class=\"iconsmall\" alt=\"$stredit\" /></a> ";
-        $buttons .= "<a title=\"$strdelete\" href=\"grouping.php?id=$grouping->id&amp;delete=1\"><img".
-                    " src=\"" . $OUTPUT->pix_url('t/delete') . "\" class=\"iconsmall\" alt=\"$strdelete\" /></a> ";
+        if (empty($grouping->idnumber) || $canchangeidnumber) {
+            // It's only possible to delete groups without an idnumber unless the user has the changeidnumber capability
+            $buttons .= "<a title=\"$strdelete\" href=\"grouping.php?id=$grouping->id&amp;delete=1\"><img".
+                        " src=\"" . $OUTPUT->pix_url('t/delete') . "\" class=\"iconsmall\" alt=\"$strdelete\" /></a> ";
+        } else {
+            $buttons .= $OUTPUT->spacer();
+        }
         $buttons .= "<a title=\"$strmanagegrping\" href=\"assign.php?id=$grouping->id\"><img".
                     " src=\"" . $OUTPUT->pix_url('i/group') . "\" class=\"icon\" alt=\"$strmanagegrping\" /></a> ";
