+++ lib/ddl/mysql_sql_generator.php
@@ -1,5 +1,4 @@
 <?php
-
 // This file is part of Moodle - http://moodle.org/
 //
 // Moodle is free software: you can redistribute it and/or modify
@@ -15,12 +14,10 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
-
 /**
  * Mysql specific SQL code generator.
  *
- * @package    core
- * @subpackage ddl
+ * @package    core_ddl
  * @copyright  1999 onwards Martin Dougiamas     http://dougiamas.com
  *             2001-3001 Eloy Lafuente (stronk7) http://contiento.com
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
@@ -30,57 +27,72 @@ defined('MOODLE_INTERNAL') || die();
 
 require_once($CFG->libdir.'/ddl/sql_generator.php');
 
-/// This class generate SQL code to be used against MySQL
-/// It extends XMLDBgenerator so everything can be
-/// overridden as needed to generate correct SQL.
-
+/**
+ * This class generate SQL code to be used against MySQL
+ * It extends XMLDBgenerator so everything can be
+ * overridden as needed to generate correct SQL.
+ *
+ * @package    core_ddl
+ * @copyright  1999 onwards Martin Dougiamas     http://dougiamas.com
+ *             2001-3001 Eloy Lafuente (stronk7) http://contiento.com
+ * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
 class mysql_sql_generator extends sql_generator {
 
-/// Only set values that are different from the defaults present in XMLDBgenerator
+    // Only set values that are different from the defaults present in XMLDBgenerator
 
-    public $quote_string = '`';   // String used to quote names
+    /** @var string Used to quote names. */
+    public $quote_string = '`';
 
-    public $default_for_char = '';      // To define the default to set for NOT NULLs CHARs without default (null=do nothing)
+    /** @var string To define the default to set for NOT NULLs CHARs without default (null=do nothing).*/
+    public $default_for_char = '';
 
-    public $drop_default_value_required = true; //To specify if the generator must use some DEFAULT clause to drop defaults
-    public $drop_default_value = NULL; //The DEFAULT clause required to drop defaults
+    /** @var bool To specify if the generator must use some DEFAULT clause to drop defaults.*/
+    public $drop_default_value_required = true;
 
-    public $primary_key_name = ''; //To force primary key names to one string (null=no force)
+    /** @var string The DEFAULT clause required to drop defaults.*/
+    public $drop_default_value = null;
 
-    public $drop_primary_key = 'ALTER TABLE TABLENAME DROP PRIMARY KEY'; // Template to drop PKs
-                // with automatic replace for TABLENAME and KEYNAME
+    /** @var string To force primary key names to one string (null=no force).*/
+    public $primary_key_name = '';
 
-    public $drop_unique_key = 'ALTER TABLE TABLENAME DROP KEY KEYNAME'; // Template to drop UKs
-                // with automatic replace for TABLENAME and KEYNAME
+    /** @var string Template to drop PKs. 'TABLENAME' and 'KEYNAME' will be replaced from this template.*/
+    public $drop_primary_key = 'ALTER TABLE TABLENAME DROP PRIMARY KEY';
 
-    public $drop_foreign_key = 'ALTER TABLE TABLENAME DROP FOREIGN KEY KEYNAME'; // Template to drop FKs
-                // with automatic replace for TABLENAME and KEYNAME
+    /** @var string Template to drop UKs. 'TABLENAME' and 'KEYNAME' will be replaced from this template.*/
+    public $drop_unique_key = 'ALTER TABLE TABLENAME DROP KEY KEYNAME';
 
-    public $sequence_extra_code = false; //Does the generator need to add extra code to generate the sequence fields
-    public $sequence_name = 'auto_increment'; //Particular name for inline sequences in this generator
+    /** @var string Template to drop FKs. 'TABLENAME' and 'KEYNAME' will be replaced from this template.*/
+    public $drop_foreign_key = 'ALTER TABLE TABLENAME DROP FOREIGN KEY KEYNAME';
 
-    public $enum_extra_code = false; //Does the generator need to add extra code to generate code for the enums in the table
+    /** @var bool True if the generator needs to add extra code to generate the sequence fields.*/
+    public $sequence_extra_code = false;
+
+    /** @var string The particular name for inline sequences in this generator.*/
+    public $sequence_name = 'auto_increment';
 
     public $add_after_clause = true; // Does the generator need to add the after clause for fields
 
-    public $concat_character = null; //Characters to be used as concatenation operator. If not defined
-                                  //MySQL CONCAT function will be use
+    /** @var string Characters to be used as concatenation operator.*/
+    public $concat_character = null;
 
-    public $alter_column_sql = 'ALTER TABLE TABLENAME MODIFY COLUMN COLUMNSPECS'; //The SQL template to alter columns
+    /** @var string The SQL template to alter columns where the 'TABLENAME' and 'COLUMNSPECS' keywords are dynamically replaced.*/
+    public $alter_column_sql = 'ALTER TABLE TABLENAME MODIFY COLUMN COLUMNSPECS';
 
-    public $drop_index_sql = 'ALTER TABLE TABLENAME DROP INDEX INDEXNAME'; //SQL sentence to drop one index
-                                                               //TABLENAME, INDEXNAME are dynamically replaced
+    /** @var string SQL sentence to drop one index where 'TABLENAME', 'INDEXNAME' keywords are dynamically replaced.*/
+    public $drop_index_sql = 'ALTER TABLE TABLENAME DROP INDEX INDEXNAME';
 
-    public $rename_index_sql = null; //SQL sentence to rename one index (MySQL doesn't support this!)
-                                      //TABLENAME, OLDINDEXNAME, NEWINDEXNAME are dynamically replaced
+    /** @var string SQL sentence to rename one index where 'TABLENAME', 'OLDINDEXNAME' and 'NEWINDEXNAME' are dynamically replaced.*/
+    public $rename_index_sql = null;
 
-    public $rename_key_sql = null; //SQL sentence to rename one key (MySQL doesn't support this!)
-                                      //TABLENAME, OLDKEYNAME, NEWKEYNAME are dynamically replaced
+    /** @var string SQL sentence to rename one key 'TABLENAME', 'OLDKEYNAME' and 'NEWKEYNAME' are dynamically replaced.*/
+    public $rename_key_sql = null;
 
     /**
      * Reset a sequence to the id field of a table.
-     * @param string $table name of table or xmldb_table object
-     * @return array sql commands to execute
+     *
+     * @param xmldb_table|string $table name of table or the table object.
+     * @return array of sql statements
      */
     public function getResetSequenceSQL($table) {
 
@@ -162,18 +174,28 @@ class mysql_sql_generator extends sql_generator {
     }
 
     /**
-     * Given one correct xmldb_table and the new name, returns the SQL statements
-     * to drop it (inside one array)
+     * Given one correct xmldb_table, returns the SQL statements
+     * to drop it (inside one array).
+     *
+     * @param xmldb_table $xmldb_table The table to drop.
+     * @return array SQL statement(s) for dropping the specified table.
      */
-    public function getDropTempTableSQL($xmldb_table) {
-        $sqlarr = $this->getDropTableSQL($xmldb_table);
-        $sqlarr = preg_replace('/^DROP TABLE/', "DROP TEMPORARY TABLE", $sqlarr);
-        $this->temptables->delete_temptable($xmldb_table->getName());
+    public function getDropTableSQL($xmldb_table) {
+        $sqlarr = parent::getDropTableSQL($xmldb_table);
+        if ($this->temptables->is_temptable($xmldb_table->getName())) {
+            $sqlarr = preg_replace('/^DROP TABLE/', "DROP TEMPORARY TABLE", $sqlarr);
+            $this->temptables->delete_temptable($xmldb_table->getName());
+        }
         return $sqlarr;
     }
 
     /**
-     * Given one XMLDB Type, length and decimals, returns the DB proper SQL type
+     * Given one XMLDB Type, length and decimals, returns the DB proper SQL type.
+     *
+     * @param int $xmldb_type The xmldb_type defined constant. XMLDB_TYPE_INTEGER and other XMLDB_TYPE_* constants.
+     * @param int $xmldb_length The length of that data type.
+     * @param int $xmldb_decimals The decimal places of precision of the data type.
+     * @return string The DB defined data type.
      */
     public function getTypeSQL($xmldb_type, $xmldb_length=null, $xmldb_decimals=null) {
 
@@ -236,16 +258,7 @@ class mysql_sql_generator extends sql_generator {
                 }
                 break;
             case XMLDB_TYPE_TEXT:
-                if (empty($xmldb_length)) {
-                    $xmldb_length = 'small';
-                }
-                if ($xmldb_length == 'small') {
-                    $dbtype = 'TEXT';
-                } else if ($xmldb_length == 'medium') {
-                    $dbtype = 'MEDIUMTEXT';
-                } else {
-                    $dbtype = 'LONGTEXT';
-                }
+                $dbtype = 'LONGTEXT';
                 if ($collation = $this->mdb->get_dbcollation()) {
                     if (strpos($collation, 'utf8_') === 0) {
                         $dbtype .= " CHARACTER SET utf8";
@@ -254,16 +267,7 @@ class mysql_sql_generator extends sql_generator {
                 }
                 break;
             case XMLDB_TYPE_BINARY:
-                if (empty($xmldb_length)) {
-                    $xmldb_length = 'small';
-                }
-                if ($xmldb_length == 'small') {
-                    $dbtype = 'BLOB';
-                } else if ($xmldb_length == 'medium') {
-                    $dbtype = 'MEDIUMBLOB';
-                } else {
-                    $dbtype = 'LONGBLOB';
-                }
+                $dbtype = 'LONGBLOB';
                 break;
             case XMLDB_TYPE_DATETIME:
                 $dbtype = 'DATETIME';
@@ -272,56 +276,41 @@ class mysql_sql_generator extends sql_generator {
     }
 
     /**
-     * Given one xmldb_table and one xmldb_field, return the SQL statements needed to create its enum
-     * (usually invoked from getModifyEnumSQL()
-     */
-    public function getCreateEnumSQL($xmldb_table, $xmldb_field) {
-    /// For MySQL, just alter the field
-        return $this->getAlterFieldSQL($xmldb_table, $xmldb_field);
-    }
-
-    /**
-     * Given one xmldb_table and one xmldb_field, return the SQL statements needed to drop its enum
-     * (usually invoked from getModifyEnumSQL()
-     *
-     * TODO: Moodle 2.1 - drop in Moodle 2.1
-     */
-    public function getDropEnumSQL($xmldb_table, $xmldb_field) {
-    /// Let's introspect to know if there is one enum
-        if ($check_constraints = $this->getCheckConstraintsFromDB($xmldb_table, $xmldb_field)) {
-        /// For MySQL, just alter the field
-            return $this->getAlterFieldSQL($xmldb_table, $xmldb_field);
-        } else {
-            return array(); /// Enum not found. Nothing to do
-        }
-    }
-
-    /**
-     * Given one xmldb_table and one xmldb_field, return the SQL statements needed to create its default
+     * Given one xmldb_table and one xmldb_field, return the SQL statements needed to add its default
      * (usually invoked from getModifyDefaultSQL()
+     *
+     * @param xmldb_table $xmldb_table The xmldb_table object instance.
+     * @param xmldb_field $xmldb_field The xmldb_field object instance.
+     * @return array Array of SQL statements to create a field's default.
      */
     public function getCreateDefaultSQL($xmldb_table, $xmldb_field) {
-    /// Just a wrapper over the getAlterFieldSQL() function for MySQL that
-    /// is capable of handling defaults
+        // Just a wrapper over the getAlterFieldSQL() function for MySQL that
+        // is capable of handling defaults
         return $this->getAlterFieldSQL($xmldb_table, $xmldb_field);
     }
 
     /**
      * Given one correct xmldb_field and the new name, returns the SQL statements
-     * to rename it (inside one array)
-     * MySQL is pretty different from the standard to justify this overloading
+     * to rename it (inside one array).
+     *
+     * @param xmldb_table $xmldb_table The table related to $xmldb_field.
+     * @param xmldb_field $xmldb_field The instance of xmldb_field to get the renamed field from.
+     * @param string $newname The new name to rename the field to.
+     * @return array The SQL statements for renaming the field.
      */
     public function getRenameFieldSQL($xmldb_table, $xmldb_field, $newname) {
+        // NOTE: MySQL is pretty different from the standard to justify this overloading.
 
-    /// Need a clone of xmldb_field to perform the change leaving original unmodified
+        // Need a clone of xmldb_field to perform the change leaving original unmodified
         $xmldb_field_clone = clone($xmldb_field);
 
-    /// Change the name of the field to perform the change
-        $xmldb_field_clone->setName($xmldb_field_clone->getName() . ' ' . $newname);
+        // Change the name of the field to perform the change
+        $xmldb_field_clone->setName($newname);
 
         $fieldsql = $this->getFieldSQL($xmldb_table, $xmldb_field_clone);
 
-        $sql = 'ALTER TABLE ' . $this->getTableName($xmldb_table) . ' CHANGE ' . $fieldsql;
+        $sql = 'ALTER TABLE ' . $this->getTableName($xmldb_table) . ' CHANGE ' .
+               $xmldb_field->getName() . ' ' . $fieldsql;
 
         return array($sql);
     }
@@ -329,15 +318,26 @@ class mysql_sql_generator extends sql_generator {
     /**
      * Given one xmldb_table and one xmldb_field, return the SQL statements needed to drop its default
      * (usually invoked from getModifyDefaultSQL()
+     *
+     * Note that this method may be dropped in future.
+     *
+     * @param xmldb_table $xmldb_table The xmldb_table object instance.
+     * @param xmldb_field $xmldb_field The xmldb_field object instance.
+     * @return array Array of SQL statements to create a field's default.
+     *
+     * @todo MDL-31147 Moodle 2.1 - Drop getDropDefaultSQL()
      */
     public function getDropDefaultSQL($xmldb_table, $xmldb_field) {
-    /// Just a wrapper over the getAlterFieldSQL() function for MySQL that
-    /// is capable of handling defaults
+        // Just a wrapper over the getAlterFieldSQL() function for MySQL that
+        // is capable of handling defaults
         return $this->getAlterFieldSQL($xmldb_table, $xmldb_field);
     }
 
     /**
-     * Returns the code (in array) needed to add one comment to the table
+     * Returns the code (array of statements) needed to add one comment to the table.
+     *
+     * @param xmldb_table $xmldb_table The xmldb_table object instance.
+     * @return array Array of SQL statements to add one comment to the table.
      */
     function getCommentSQL ($xmldb_table) {
         $comment = '';
@@ -350,78 +350,29 @@ class mysql_sql_generator extends sql_generator {
     }
 
     /**
-     * Given one xmldb_table returns one array with all the check constraints
-     * in the table (fetched from DB)
-     * Optionally the function allows one xmldb_field to be specified in
-     * order to return only the check constraints belonging to one field.
-     * Each element contains the name of the constraint and its description
-     * If no check constraints are found, returns an empty array
-     * MySQL doesn't have check constraints in this implementation, but
-     * we return them based on the enum fields in the table
+     * Given one object name and it's type (pk, uk, fk, ck, ix, uix, seq, trg).
      *
-     * TODO: Moodle 2.1 - drop in Moodle 2.1
-     */
-    public function getCheckConstraintsFromDB($xmldb_table, $xmldb_field = null) {
-
-        $tablename = $xmldb_table->getName($xmldb_table);
-
-    /// Fetch all the columns in the table
-        if (!$columns = $this->mdb->get_columns($tablename)) {
-            return array();
-        }
-
-    /// Filter by the required field if specified
-        if ($xmldb_field) {
-            $filter = $xmldb_field->getName();
-            if (!isset($columns[$filter])) {
-                return array();
-            }
-            $column = ($columns[$filter]);
-            if (!empty($column->enums)) {
-                $result = new stdClass();
-                $result->name = $filter;
-                $result->description = implode(', ', $column->enums);
-                return array($result);
-            } else {
-                return array();
-            }
-
-        } else {
-            $results = array();
-        /// Iterate over columns searching for enums
-            foreach ($columns as $key => $column) {
-            /// Enum found, let's add it to the constraints list
-                if (!empty($column->enums)) {
-                    $result = new stdClass();
-                    $result->name = $key;
-                    $result->description = implode(', ', $column->enums);
-                    $results[$key] = $result;
-                }
-            }
-            return $results;
-        }
-    }
-
-    /**
-     * Given one object name and it's type (pk, uk, fk, ck, ix, uix, seq, trg)
-     * return if such name is currently in use (true) or no (false)
-     * (invoked from getNameForObject()
+     * (MySQL requires the whole xmldb_table object to be specified, so we add it always)
+     *
+     * This is invoked from getNameForObject().
+     * Only some DB have this implemented.
+     *
+     * @param string $object_name The object's name to check for.
+     * @param string $type The object's type (pk, uk, fk, ck, ix, uix, seq, trg).
+     * @param string $table_name The table's name to check in
+     * @return bool If such name is currently in use (true) or no (false)
      */
     public function isNameInUse($object_name, $type, $table_name) {
-
-    /// Calculate the real table name
-        $xmldb_table = new xmldb_table($table_name);
-        $tname = $this->getTableName($xmldb_table);
 
         switch($type) {
             case 'ix':
             case 'uix':
-            /// First of all, check table exists
+                // First of all, check table exists
                 $metatables = $this->mdb->get_tables();
-                if (isset($metatables[$tname])) {
-                /// Fetch all the indexes in the table
-                    if ($indexes = $this->mdb->get_indexes($tname)) {
-                    /// Look for existing index in array
+                if (isset($metatables[$table_name])) {
+                    // Fetch all the indexes in the table
+                    if ($indexes = $this->mdb->get_indexes($table_name)) {
+                        // Look for existing index in array
                         if (isset($indexes[$object_name])) {
                             return true;
                         }
@@ -435,10 +386,11 @@ class mysql_sql_generator extends sql_generator {
 
     /**
      * Returns an array of reserved words (lowercase) for this DB
+     * @return array An array of database specific reserved words
      */
     public static function getReservedWords() {
-    /// This file contains the reserved words for MySQL databases
-    /// from http://dev.mysql.com/doc/refman/6.0/en/reserved-words.html
+        // This file contains the reserved words for MySQL databases
+        // from http://dev.mysql.com/doc/refman/6.0/en/reserved-words.html
         $reserved_words = array (
             'accessible', 'add', 'all', 'alter', 'analyze', 'and', 'as', 'asc',
             'asensitive', 'before', 'between', 'bigint', 'binary',
